rename: markAsRead -> markAsReadScheduler
