+++ src/lib/Sendbird.tsx
@@ -25,7 +25,7 @@ import { LocalizationProvider } from './LocalizationContext';
 import { MediaQueryProvider } from './MediaQueryContext';
 import getStringSet from '../ui/Label/stringSet';
 import { VOICE_RECORDER_DEFAULT_MAX, VOICE_RECORDER_DEFAULT_MIN } from '../utils/consts';
-import { useMarkAsRead } from './hooks/useMarkAsRead';
+import { useMarkAsReadScheduler } from './hooks/useMarkAsReadScheduler';
 
 export type UserListQueryType = {
   hasNext?: boolean;
@@ -198,7 +198,7 @@ const Sendbird = ({
 
   const isOnline = useOnlineStatus(sdkStore.sdk, logger);
 
-  const markAsRead = useMarkAsRead({ isConnected: isOnline }, { logger });
+  const markAsReadScheduler = useMarkAsReadScheduler({ isConnected: isOnline }, { logger });
 
   const localeStringSet = React.useMemo(() => {
     if (!stringSet) {
@@ -266,7 +266,7 @@ const Sendbird = ({
           isTypingIndicatorEnabledOnChannelList,
           isMessageReceiptStatusEnabledOnChannelList,
           replyType,
-          markAsRead,
+          markAsReadScheduler,
         },
       }}
     >
