+++ src/lib/SendbirdState.tsx
@@ -2,7 +2,7 @@ import type SendbirdChat from '@sendbird/chat';
 import type { User } from '@sendbird/chat';
 import type { GroupChannel, Member } from '@sendbird/chat/groupChannel';
 import { Dispatch } from 'react';
-import { MarkAsReadSchedulerType } from './hooks/useMarkAsRead';
+import { MarkAsReadSchedulerType } from './hooks/useMarkAsReadScheduler';
 
 export type CustomUseReducerDispatcher = Dispatch<{ type: string, payload: any }>;
